+++ xmake.lua
@@ -9,7 +9,6 @@ add_repositories("OTOTYAN https://github.com/OEOTYAN/xmake-repo.git")
 -- LeviMc(LiteLDev)
 add_requires("levilamina 1.3.1", {configs = {target_type = "server"}})
 add_requires("levibuildscript 0.4.0")
-add_requires("legacymoney 0.9.0-rc.1")
 
 -- OTOTYAN
 add_requires("bsci 0.1.6")
@@ -70,7 +69,6 @@ target("PLand") -- Change this to your mod name.
         "levilamina",
         "exprtk",
         "ilistenattentively",
-        "legacymoney",
         "bsci"
     )